pyproject.toml
[project]
name = "g-track-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "joblib>=1.5.1",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
